projects/spotifly/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
croniter==1.3.7
Deprecated==1.2.13
Django==4.1.2
django-model-utils==4.2.0
git+https://github.com/silviolleite/django-pwa
django-redis==5.2.0
django-rq==2.5.1
gunicorn==20.1.0
idna==3.4
packaging==21.3
psycopg2-binary==2.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.5
redis==4.3.4
requests==2.28.1
rq==1.11.1
rq-scheduler==0.11.0
six==1.16.0
spotipy==2.20.0
sqlparse==0.4.3
urllib3==1.26.12
wrapt==1.14.1
